ServiceProblem Searching now works with superman-core. Started Custom Serializers

+++ bango-test/src/main/java/sonique/bango/driver/AppContainer.java
@@ -20,7 +20,7 @@ public class AppContainer extends SupermanContainer {
         return new SearchPanel(this);
     }
 
-    public ServiceProblemTab serviceProblemTab(Integer serviceProblemId) {
+    public ServiceProblemTab serviceProblemTab(Long serviceProblemId) {
         return new ServiceProblemTab(this, serviceProblemId);
     }
 }
