+++ bango-test/src/main/java/sonique/bango/driver/panel/ServiceProblemTab.java
@@ -6,7 +6,7 @@ import org.openqa.selenium.WebElement;
 import static org.openqa.selenium.By.id;
 
 public class ServiceProblemTab extends SupermanComponent implements HasTitle {
-    public ServiceProblemTab(SupermanContainer container, Integer serviceProblemId) {
+    public ServiceProblemTab(SupermanContainer container, Long serviceProblemId) {
         super(container, id(String.format("service-problem-tab-%d", serviceProblemId)));
     }
